requirements.txt
numpy>=1.24.0,<2.0
scipy>=1.10.0
pandas>=2.0.0
scikit-surprise>=1.1.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
